FloatField en vez de DecimalField

+++ django_backend/apps/maestros/models.py
@@ -10,7 +10,7 @@ class Producto(models.Model):
 	"""
 	codigo = models.CharField(max_length=50)
 	descripcion = models.TextField(null=True, blank=True)
-	precio_venta = models.DecimalField (max_digits=4, decimal_places=2, default=0)
+	precio_venta = models.FloatField (default=0)
 	esta_activo  = models.BooleanField(default=True)
 
 	registered_at = models.DateTimeField(auto_now_add=True)
